Avoid redefinitions of JIT'ed functions in repeated calls (#14)
* Move functions  to avoid redefinitions, do not JIT outer simulation function

JIT'ing run_simulation is not really necessary, as we generally only simulate a network once. If the network/args change, this  function would be recompiled anyway. Thus, its easier to save the time and effor to JIT this function. This might in the future if more of the network information would be stored in the state variable instead of as class attributes, avoiding the need to recompile when the network changes.

Furthermore, the inner functions were being redefined every time run_simulation was called. Moving them outside avoids this redefinition overhead, and allows them to be reused across multiple calls to run_simulation.

* remove print statements