Allow None noise shape when no noise is needed (#20)
* Add DefaultIfNone operator

* Do not generate noise in LIF model, use appropriate diffusion term.

Instead of generating noise which is multiplied by zero, we now generate only None values for the noise (way cheaper). The diffusion term is now defined using the DefaultIfNone operator, which returns a default value (zeros of the right shape) if the noise is None. This way, the stochastic term still generates the same (namely all zeros of the correct shape), but without the overhead of generating noise that is not used. This significantly reduces runtime (about 50% from my tests).

* Don't generate noise for reward model

* Don't generate noise in noiseless envs

* Add DI env with kick control

* Update tests